+++ changelog.md
@@ -1,3 +1,7 @@
+## April 17th 2019 -
+
+- markRead now supports sending a message_id to mark the channel read up to (and including) that specific message
+
 ## April 9th 2019 - 0.7.2
 
 - queryChannels used to return the list of members twice, this has now been resolved. However if you were using the duplicate list of members in channel.members you'll want to update to Object.values(channel.state.members)